Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -9,7 +9,7 @@ int main(){
     scanf("%d",&b);
     scanf("%d",&c);
     sum=a+b+c;
-    avg=sum/3;
+    avg=(float)sum/3;
     printf("Average: %.2f",avg);
     return 0;
 }
